+++ src/onegov.town/HISTORY.rst
@@ -1,6 +1,9 @@
 Changelog
 ---------
 
+- Fixes failing tests.
+  [msom]
+
 1.13.1 (2017-06-16)
 ~~~~~~~~~~~~~~~~~~~~~